Refactored URL params into own hook
