+++ src/components/identity-registrar.tsx
@@ -34,6 +34,7 @@ import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, D
 import { config } from "~/api/config"
 import TeleporterDialog from "./dialogs/Teleporter"
 import { decodeAddress, encodeAddress } from "@polkadot/keyring";
+import { useUrlParams } from "~/hooks/useUrlParams"
 
 export function IdentityRegistrarComponent() {
   const [currentPage, setCurrentPage] = useState(0)
@@ -49,26 +50,7 @@ export function IdentityRegistrarComponent() {
 
   const accountStore = useProxy(_accountStore)
 
-  const urlParams = useMemo<{
-    chain: string;
-    address: string;
-  }>(() => {
-    const _urlParams = new URLSearchParams(window.location.search).entries()
-      .map(([key, value]) => ({ key, value }))
-      .reduce((acc, { key, value }) => ({ ...acc, [key]: value }), {})
-    if (import.meta.env.DEV) console.log({ _urlParams })
-    return _urlParams
-  }, [window.location.search])
-  const updateUrlParams = useCallback((params) => {
-    const newParams = params
-      ? "?" + Object.entries(params)
-        .filter(([ , value ]) => value)
-        .map(([ key, value ]) => `${key}=${value}`)
-        .join("&")
-      : ""
-    window.history.replaceState(null, null, `${window.location.pathname}${newParams}`)
-    if (import.meta.env.DEV) console.log({ newParams })
-  }, [window.history.replaceState])
+  const { urlParams, updateUrlParams } = useUrlParams()
 
   const pages = [
     {
